refactor: split manager helpers into separate files
