+++ .sonarcloud.properties
@@ -1,1 +1,1 @@
-sonar.cpd.exclusions=**/*_test.go,**/testabilities/**,**/fixtures/**,**/tools/**,**/examples/**,**/mocks/**
+sonar.cpd.exclusions=**/*_test.go,**/testabilities/**,**/fixtures/**,**/tools/**,**/examples/**,**/mocks/**, **/manual_tests/**
