+++ manual_tests/internal/tui/balance_view.go
@@ -2,6 +2,7 @@ package tui
 
 import (
 	"fmt"
+	
 	"github.com/bsv-blockchain/go-wallet-toolbox-manual-tests/internal/fixtures"
 	"github.com/charmbracelet/bubbles/spinner"
 	tea "github.com/charmbracelet/bubbletea"
