+++ manual_tests/internal/tui/interfaces.go
@@ -2,6 +2,7 @@ package tui
 
 import (
 	"context"
+
 	"github.com/bsv-blockchain/go-wallet-toolbox-manual-tests/internal/fixtures"
 	"github.com/bsv-blockchain/go-wallet-toolbox/pkg/brc29"
 	"github.com/bsv-blockchain/go-wallet-toolbox/pkg/defs"
