+++ manual_tests/internal/utils.go
@@ -8,7 +8,7 @@ import (
 func BytesFromBase64(s string) ([]byte, error) {
 	result, err := base64.StdEncoding.DecodeString(s)
 	if err != nil {
-		return nil, fmt.Errorf("failed to decode base64 input: %s", err)
+		return nil, fmt.Errorf("failed to decode base64 input: %w", err)
 	}
-	return result, err
+	return result, nil
 }
